+++ src/pages/home/Home.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta } from "@storybook/react";
-import { Home } from "./Home";
+import { Home } from "./index.page";
 
 export default {
   title: "pages/home"
