--- src/pages/home/Home.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import * as Styled from "./Home.styles";
-import { SelectionProps } from "./Home.types";
-
-export const Selection = (ContainerProps: SelectionProps) => {
-  const content = (
-    <Styled.SelectBox>
-      <Styled.SelectButton>
-        <Styled.SelectIcon svgName={ContainerProps.svgName} />
-      </Styled.SelectButton>
-      <Styled.SelectText>{ContainerProps.name}</Styled.SelectText>
-    </Styled.SelectBox>
-  );
-  return content;
-};
